Display registered authenticators
